Remove false positives in CNameRecord.py and minor clean up

+++ Cerebrum/modules/dns/CNameRecord.py
@@ -3,6 +3,7 @@
 from Cerebrum.Entity import Entity
 from Cerebrum.modules.dns.DnsOwner import DnsOwner
 
+
 class CNameRecord(Entity):
     """``CNameRecord.CNameRecord(Entity)`` is used to
     store information about CName-records in the dns_cname_record
@@ -28,7 +29,7 @@ class CNameRecord(Entity):
             Entity.populate(self, self.const.entity_dns_cname)
         try:
             if not self.__in_db:
-                raise RuntimeError, "populate() called multiple times."
+                raise RuntimeError("populate() called multiple times.")
         except AttributeError:
             self.__in_db = False
         for k in locals().keys():
@@ -37,9 +38,10 @@ class CNameRecord(Entity):
 
     def __eq__(self, other):
         assert isinstance(other, CNameRecord)
-        if (self.cname_owner_id != other.cname_owner_id or
-            self.ttl != other.ttl or 
-            self.target_owner_id != other.target_owner_id):
+        if (
+                self.cname_owner_id != other.cname_owner_id or
+                self.ttl != other.ttl or
+                self.target_owner_id != other.target_owner_id):
             return False
         return True
 
@@ -48,45 +50,55 @@ class CNameRecord(Entity):
         if not self.__updated:
             return
         is_new = not self.__in_db
-        cols = [('entity_type', ':e_type'),
-                ('cname_id', ':e_id'),
-                ('cname_owner_id', ':cname_owner_id'),
-                ('ttl', ':ttl'),
-                ('target_owner_id', ':target_owner_id')]
-        binds = {'e_type' : int(self.const.entity_dns_cname),
-                 'e_id': self.entity_id,
+        binds = {'entity_type': int(self.const.entity_dns_cname),
+                 'cname_id': self.entity_id,
                  'cname_owner_id': self.cname_owner_id,
-                 'ttl' : self.ttl,
-                 'target_owner_id' : self.target_owner_id}
+                 'ttl': self.ttl,
+                 'target_owner_id': self.target_owner_id}
+        defs = {'tc': ', '.join(x for x in binds),
+                'tb': ', '.join(':{0}'.format(x) for x in binds),
+                'ts': ', '.join('{0}=:{0}'.format(x) for x in binds
+                                if x != 'cname_id'),
+                'tw': ' AND '.join('{0}=:{0}'.format(x) for x in binds)}
         if is_new:
-            self.execute("""
-            INSERT INTO [:table schema=cerebrum name=dns_cname_record] (%(tcols)s)
-            VALUES (%(binds)s)""" % {'tcols': ", ".join([x[0] for x in cols]),
-                                     'binds': ", ".join([x[1] for x in cols])},
-                         binds)
+            insert_stmt = """
+            INSERT INTO [:table schema=cerebrum name=dns_cname_record] (%(tc)s)
+            VALUES (%(tb)s)""" % defs
+            self.execute(insert_stmt, binds)
             self._db.log_change(self.cname_owner_id, self.clconst.cname_add,
                                 self.target_owner_id)
         else:
-            self.execute("""
-            UPDATE [:table schema=cerebrum name=dns_cname_record]
-            SET %(defs)s
-            WHERE cname_id=:e_id""" % {'defs': ", ".join(
-                ["%s=%s" % x for x in cols])},
-                         binds)
-            self._db.log_change(self.cname_owner_id, self.clconst.cname_update,
-                                self.target_owner_id)
+            exists_stmt = """
+              SELECT EXISTS (
+                SELECT 1
+                FROM [:table schema=cerebrum name=dns_cname_record]
+                WHERE %(tw)s
+              )
+            """ % defs
+            if not self.query_1(exists_stmt, binds):
+                # True positive
+                update_stmt = """
+                UPDATE [:table schema=cerebrum name=dns_cname_record]
+                SET %(ts)s
+                WHERE cname_id=:cname_id""" % defs
+                self.execute(update_stmt, binds)
+                self._db.log_change(self.cname_owner_id,
+                                    self.clconst.cname_update,
+                                    self.target_owner_id)
         del self.__in_db
         self.__in_db = True
         self.__updated = []
         return is_new
 
     def find(self, cname_id):
+        """Find by cname_id"""
         self.__super.find(cname_id)
-        (self.cname_owner_id, self.ttl, self.target_owner_id
-         ) = self.query_1("""
-        SELECT cname_owner_id, ttl, target_owner_id
-        FROM [:table schema=cerebrum name=dns_cname_record]
-        WHERE cname_id=:cname_id""", {'cname_id' : cname_id})
+        (self.cname_owner_id,
+         self.ttl,
+         self.target_owner_id) = self.query_1("""
+         SELECT cname_owner_id, ttl, target_owner_id
+         FROM [:table schema=cerebrum name=dns_cname_record]
+         WHERE cname_id=:cname_id""", {'cname_id': cname_id})
         dns = DnsOwner(self._db)
         dns.find(self.cname_owner_id)
         self.name = dns.name
@@ -98,15 +110,18 @@ class CNameRecord(Entity):
         self.__updated = []
 
     def find_by_name(self, name):
+        """Find by name"""
         dns = DnsOwner(self._db)
         dns.find_by_name(name)
         self.find_by_cname_owner_id(dns.entity_id)
 
     def find_by_cname_owner_id(self, cname_owner_id):
+        """Find by cname_owner_id"""
         cname_id = self.query_1("""
         SELECT cname_id
         FROM [:table schema=cerebrum name=dns_cname_record]
-        WHERE cname_owner_id=:cname_owner_id""", {'cname_owner_id': cname_owner_id})
+        WHERE cname_owner_id=:cname_owner_id""",
+                                {'cname_owner_id': cname_owner_id})
         self.find(cname_id)
 
     def list_ext(self, target_owner=None, cname_owner=None, zone=None):
@@ -117,7 +132,7 @@ class CNameRecord(Entity):
         if target_owner:
             where.append("c.target_owner_id=:target_owner_id")
         if cname_owner:
-            where.append("c.cname_owner_id=:cname_owner_id")            
+            where.append("c.cname_owner_id=:cname_owner_id")
         if zone is not None:
             where.append("d_own.zone_id=:zone")
             zone = int(zone)
@@ -133,15 +148,28 @@ class CNameRecord(Entity):
         WHERE %s ORDER BY en_own.entity_name""" % where, {
             'target_owner_id': target_owner,
             'cname_owner_id': cname_owner,
-            'zone': zone} )
+            'zone': zone})
 
     def _delete(self):
         """Deletion in cname_record should be done through the DnsHelper
         class to avoid leaving entries in dns_owner that has no FKs to
         it"""
-        self.execute("""
-        DELETE FROM [:table schema=cerebrum name=dns_cname_record]
-        WHERE cname_id=:e_id""", {'e_id': self.entity_id})
+        binds = {'e_id': self.entity_id}
+        exists_stmt = """
+          SELECT EXISTS (
+            SELECT 1
+            FROM [:table schema=cerebrum name=dns_cname_record]
+            WHERE cname_id=:e_id
+          )
+        """
+        if not self.query_1(exists_stmt, binds):
+            # False positive
+            return
+        delete_stmt = """
+          DELETE FROM [:table schema=cerebrum name=dns_cname_record]
+          WHERE cname_id=:e_id
+        """
+        self.execute(delete_stmt, binds)
         self._db.log_change(self.cname_owner_id, self.clconst.cname_del,
                             self.target_owner_id)
         self.__super.delete()
